Add files via upload
This is the file resulted from the first task of data collection, in csv format. It is a dataset of 1100 x 93 where the rows represent the single artworks (1100) and the columns represent the variables (93) as retrieved from the API. The meaning of the variables and their type can be found at the following link: https://api.artic.edu/docs/#collections-2. We would suggest to carefully inspect the dataset and the documentation before carrying out any analyses.